tcsite/requirements.txt
Django==1.9.5
Pillow==3.2.0
django-admin-sortable==2.0.15
django-modeladmin-reorder==0.1.3
psycopg2==2.6.1
